move tests to individual files
