+++ spec/basic/basic_spec_01.expected
@@ -1,0 +1,1 @@
+./SPEC/BASIC/BASIC_SPEC.RB@@PASSED@@adds two numbers together
