+++ spec/basic/basic_spec_02.expected
@@ -1,0 +1,1 @@
+./SPEC/BASIC/BASIC_SPEC.RB@@PASSED@@subtracts two numbers together
