+++ spec/basic/basic_spec_03.expected
@@ -1,0 +1,2 @@
+./SPEC/BASIC/BASIC_SPEC.RB@@PASSED@@adds two numbers together
+./SPEC/BASIC/BASIC_SPEC.RB@@PASSED@@subtracts two numbers together
