--- spec/basic_spec/basic_spec_01.expected
@@ -1,1 +1,0 @@
-./SPEC/BASIC_SPEC/BASIC_SPEC.RB@@PASSED@@adds two numbers together
