--- spec/basic_spec/basic_spec_02.expected
@@ -1,1 +1,0 @@
-./SPEC/BASIC_SPEC/BASIC_SPEC.RB@@PASSED@@subtracts two numbers together
