--- spec/basic_spec/basic_spec_03.expected
@@ -1,2 +1,0 @@
-./SPEC/BASIC_SPEC/BASIC_SPEC.RB@@PASSED@@adds two numbers together
-./SPEC/BASIC_SPEC/BASIC_SPEC.RB@@PASSED@@subtracts two numbers together
