+++ spec/blockless_it/blockless_it_spec_01.commands
@@ -1,0 +1,2 @@
+:6
+:lua require("neotest").run.run()
