+++ tests/unit/basic_spec.lua
@@ -1,0 +1,3 @@
+local test = require("tests.testing_template")
+
+test.describe("Testing basic specs", "basic_spec.rb")
