+++ tests/unit/blockless_it_spec.lua
@@ -1,0 +1,3 @@
+local test = require("tests.testing_template")
+
+test.describe("Testing blockless its", "blockless_it_spec.rb")
